+++ CHANGELOG.md
@@ -1,3 +1,22 @@
 # Changelog #
-1.0.0 - Initial version
-1.0.1 - Fix UA-XXX validation
+* 1.1.0 * 
+- Fix General Settings
+- Add Settings tab
+- Add Contact Form 7 tab
+
+* 1.0.4 *
+- Fix General settings
+- Allways publish general settings
+- Fix minor bug
+
+* 1.0.3 *
+- Fix GTM- validation + Brazilian language
+
+* 1.0.2 *
+- Update version
+
+* 1.0.1 *
+- Fix UA- validation
+
+* 1.0.0 *
+- Initial version